requirements.txt
Click==7.0
cycler==0.10.0
dnspython==1.16.0
eventlet==0.25.2
Flask==1.1.1
Flask-SocketIO==4.3.0
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
monotonic==1.5
numpy==1.18.3
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-engineio==3.12.1
python-socketio==4.5.1
six==1.14.0
Werkzeug==1.0.0
